Fixing up deployer build script for the merged installer

+++ images/gestalt-deployer-image/build.sh
@@ -8,6 +8,8 @@ REGISTRY="gcr.io/galacticfog-public"
 LABEL="gestalt-deployer"
 DEFAULT_TAG="testing"
 PRE_BUILD_SCRIPT="./pre-build.sh"
+# BUILD_OUTPUT_FILE="buildoutput"
+RM_BUILDER_IMAGE=0
 declare -a TAGS
 declare -a BUILD_ARGS
 
@@ -129,6 +131,7 @@ if [ ${#TAGS[@]} -gt 0 ]; then
 else
   debug "Building only the default tag '${DEFAULT_TAG}'"
   #PUBLISH=0
+  TAGS=( "$DEFAULT_TAG" )
 fi
 
 NOT_STRING="NOT "
@@ -146,15 +149,16 @@ get_output() {
 }
 
 if [ -z ${PRE_BUILD_SCRIPT:+x} ]; then
+  debug "No pre-build script defined.  Skipping pre-build..."
+else
   if [ -f ${PRE_BUILD_SCRIPT} ]; then
     debug "Running pre-build script ${PRE_BUILD_SCRIPT}"
-    . $PRE_BUILD_SCRIPT
-    exit_on_error "FAILED while running pre-build script ${PRE_BUILD_SCRIPT}"
+    OUTPUT=$( $PRE_BUILD_SCRIPT )
+    [ $? -eq 0 ] || exit_with_error "FAILED while running pre-build script ${PRE_BUILD_SCRIPT} $(get_output)"
+    debug "${PRE_BUILD_SCRIPT} output was $(get_output)"
   else
     error "Defined pre-build script ${PRE_BUILD_SCRIPT} was not found!  Skipping..."
   fi
-else
-  debug "No pre-build script defined.  Skipping pre-build..."
 fi
 
 #Build the image
@@ -163,9 +167,9 @@ BUILD_CMD="docker build -t ${REGISTRY}/${LABEL}:${DEFAULT_TAG} ."
 for arg in ${BUILD_ARGS[@]}; do
   BUILD_CMD="${BUILD_CMD} --build-arg $arg"
 done
-BUILD_CMD="${BUILD_CMD} --build-arg component_label=$LABEL"
+# BUILD_CMD="${BUILD_CMD} --build-arg component_label=$LABEL"
 debug "Building with command '$BUILD_CMD'"
-OUTPUT=$($BUILD_CMD 2>&1)
+OUTPUT=$($BUILD_CMD 2>&1 | tee $BUILD_OUTPUT_FILE)
 [ $? -eq 0 ] || exit_with_error "FAILED image build for '$LABEL' using command '$BUILD_CMD' $(get_output)"
 debug "$(get_output)"
 
@@ -176,18 +180,20 @@ fi
 info "----- Successfully built ${LABEL} image with ID '$imageid'"
 [ $PRINT_IMAGE_ID -eq 0 ] || echo "$imageid"
 
-FIND_BUILDER_IMAGE_CMD="docker image ls --filter=label=build.phase=builder --filter=label=component=$LABEL --filter=dangling=true --format={{.ID}}"
-debug "Searching for builder image with command '$FIND_BUILDER_IMAGE_CMD'"
-builder_imageid=$($FIND_BUILDER_IMAGE_CMD)
-if [ $? -eq 0 ]; then
-  RM_BUILDER_IMAGE_CMD="docker image rm $builder_imageid"
-  debug "Removing builder image with command '$FIND_BUILDER_IMAGE_CMD'"
-  OUTPUT=$($RM_BUILDER_IMAGE_CMD)
-  [ $? -eq 0 ] || error "Unable to remove builder image $builder_imageid with command '$RM_BUILDER_IMAGE_CMD'"
-  debug "$(get_output)"
-  info "Removed builder image $builder_imageid"
-else
-  error "Unable to search for builder image $builder_imageid with command '$FIND_BUILDER_IMAGE_CMD'"
+if [ ${RM_BUILDER_IMAGE} -ne 0 ]; then
+  FIND_BUILDER_IMAGE_CMD="docker image ls --filter=label=build.phase=builder --filter=label=component=$LABEL --filter=dangling=true --format={{.ID}}"
+  debug "Searching for builder image with command '$FIND_BUILDER_IMAGE_CMD'"
+  builder_imageid=$($FIND_BUILDER_IMAGE_CMD)
+  if [ $? -eq 0 ]; then
+    RM_BUILDER_IMAGE_CMD="docker image rm $builder_imageid"
+    debug "Removing builder image with command '$RM_BUILDER_IMAGE_CMD'"
+    OUTPUT=$($RM_BUILDER_IMAGE_CMD)
+    [ $? -eq 0 ] || error "Unable to remove builder image $builder_imageid with command '$RM_BUILDER_IMAGE_CMD'"
+    debug "$(get_output)"
+    info "Removed builder image $builder_imageid"
+  else
+    error "Unable to search for builder image $builder_imageid with command '$FIND_BUILDER_IMAGE_CMD'"
+  fi
 fi
 
 #Tag and Push
